+++ packages/node-utils/src/index.ts
@@ -1,1 +1,2 @@
+export { getFromHttps } from "./curl.js"
 export { readFileAsJsonSync, writeJsonAsFileSync } from "./json.js"